office won't install if it is already installed

+++ softwareInstallation/installPrograms.ps1
@@ -1,25 +1,35 @@
 # == Install office ==
-# Check if the Office installation directory
-# already exists and delete it if so
-if(Test-Path -Path "Office"){
+# Install office if office is not installed
+$isOfficeInstalled = Get-ItemProperty HKLM:\Software\Microsoft\Windows\CurrentVersion\Uninstall\O365ProPlusRetail* | Select-Object DisplayName,DisplayVersion,Publisher
+if($isOfficeInstalled -eq $null){
+  # Print office installing output
+  Write-Output("Office is not installed. Installing office...")
+  # Check if the Office installation directory
+  # already exists and delete it if so
+  if(Test-Path -Path "Office"){
+    Remove-Item Office -Recurse
+  }
+  # Install the corresponding office version (64 bit
+  # for 64 bit machines and 32 bit for 32 bit machines)
+  if([System.Environment]::Is64BitOperatingSystem){
+    Write-Output("Installing office 64bit...")
+    .\ms-office\setup.exe /download ms-office\configuration-Office365-x64.xml
+    .\ms-office\setup.exe /configure ms-office\configuration-Office365-x64.xml
+    Write-Output("Finished office 64bit installation")
+  }
+  else{
+    Write-Output("Installing office 32bit...")
+    .\ms-office\setup.exe /download ms-office\configuration-Office365-x86.xml
+    .\ms-office\setup.exe /configure ms-office\configuration-Office365-x86.xml
+    Write-Output("Finished office 32bit installation")
+  }
+  # Remove residue of the office installation directory
   Remove-Item Office -Recurse
 }
-# Install the corresponding office version (64 bit
-# for 64 bit machines and 32 bit for 32 bit machines)
-if([System.Environment]::Is64BitOperatingSystem){
-  Write-Output("Installing office 64bit...")
-  .\ms-office\setup.exe /download ms-office\configuration-Office365-x64.xml
-  .\ms-office\setup.exe /configure ms-office\configuration-Office365-x64.xml
-  Write-Output("Finished office 64bit installation")
-}
 else{
-  Write-Output("Installing office 32bit...")
-  .\ms-office\setup.exe /download ms-office\configuration-Office365-x86.xml
-  .\ms-office\setup.exe /configure ms-office\configuration-Office365-x86.xml
-  Write-Output("Finished office 32bit installation")
+  # Print office already installed output
+  Write-Output("Office is already installed. Proceeding...")
 }
-# Remove residue of the office installation directory
-Remove-Item Office -Recurse
 
 # == Install other applications ==
 # Rerun this script as administrator if it is not already
